server/app/requirements.txt
flask
sqlalchemy
psycopg2-binary
flask_sqlalchemy
selenium
apscheduler
webdriver-manager
schedule
flask-cors